03enWSL/pyproject.toml
# ═══════════════════════════════════════════════════════════════════════════════
# pyproject.toml — Week 3 Network Programming Laboratory Kit
# ═══════════════════════════════════════════════════════════════════════════════
# NETWORKING class - ASE, CSIE | by ing. dr. Antonio Clim
# ═══════════════════════════════════════════════════════════════════════════════

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "week3-network-programming"
version = "2.0.0"
description = "Week 3 Laboratory Kit: Network Programming with UDP Broadcast, Multicast and TCP Tunnelling"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "ing. dr. Antonio Clim", email = "antonio.clim@csie.ase.ro"}
]
keywords = ["networking", "education", "sockets", "broadcast", "multicast", "tcp", "tunnel"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
    "Topic :: System :: Networking",
]

dependencies = [
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.2.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
    "jsonschema>=4.20.0",
]

[project.scripts]
run-quiz = "formative.run_quiz:main"

# ═══════════════════════════════════════════════════════════════════════════════
# RUFF CONFIGURATION
# ═══════════════════════════════════════════════════════════════════════════════

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    ".pytest_cache",
    "htmlcov",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
    "D",      # pydocstyle (docstrings)
]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "B008",   # Function call in default argument (intentional for CLI)
    "ARG001", # Unused function argument (needed for callbacks)
    "SIM108", # Ternary instead of if-else (readability preference)
    "PTH123", # open() vs Path.open() (educational clarity)
    "D100",   # Missing docstring in public module
    "D104",   # Missing docstring in public package
    "D203",   # one-blank-line-before-class (conflicts with D211)
    "D213",   # multi-line-summary-second-line (conflicts with D212)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "ARG", "S101"]  # Allow missing docstrings in tests
"homework/*" = ["F841", "ARG", "D"]  # Allow unused variables (TODOs for students)
"formative/*" = ["D"]  # Quiz runner has inline comments

[tool.ruff.lint.isort]
known-first-party = ["src", "scripts", "formative", "anti_ai"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

# ═══════════════════════════════════════════════════════════════════════════════
# MYPY CONFIGURATION
# ═══════════════════════════════════════════════════════════════════════════════

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_equality = true
extra_checks = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
show_error_codes = true
show_column_numbers = true

[[tool.mypy.overrides]]
module = [
    "yaml.*",
    "pytest.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "tests.*",
    "homework.*",
]
disallow_untyped_defs = false
check_untyped_defs = true

# ═══════════════════════════════════════════════════════════════════════════════
# PYTEST CONFIGURATION
# ═══════════════════════════════════════════════════════════════════════════════

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--tb=short",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "docker: marks tests that require Docker containers",
    "integration: marks integration tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]
timeout = 60

# ═══════════════════════════════════════════════════════════════════════════════
# COVERAGE CONFIGURATION
# ═══════════════════════════════════════════════════════════════════════════════

[tool.coverage.run]
source = ["src", "scripts", "formative"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/homework/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
fail_under = 70
show_missing = true
skip_covered = true

[tool.coverage.html]
directory = "htmlcov"

# ═══════════════════════════════════════════════════════════════════════════════
# EOF
# ═══════════════════════════════════════════════════════════════════════════════
